+++ packages/core/src/lib/Table/MockData.tsx
@@ -2,9 +2,11 @@ import * as React from 'react';
 import { color } from '@web3uikit/styles';
 import { Avatar } from '../Avatar';
 import { Button } from '../Button';
-import { Cross, MoreVert } from '@web3uikit/icons';
+import { Cross, MoreVert, Network, Server, Testnet } from '@web3uikit/icons';
 import { Tag } from '../Tag';
 import { Typography } from '../Typography';
+import { PopoverDropdown } from '../PopoverDropdown';
+import { PopoverElement } from '../PopoverElement';
 
 export const columnsConfig = '80px 3fr 2fr 2fr 80px';
 export const header = [
@@ -559,5 +561,224 @@ export const syncData = [
                 icon={<Cross fontSize={18} />}
             />
         </>,
+    ],
+];
+
+export const customTableMockData = [
+    [
+        <div style={{ padding: '11px' }}>
+            <Avatar theme="image" isRounded={true} size={36} />
+        </div>,
+        <div style={{ padding: '11px' }}>Moralis Magi</div>,
+        <div>
+            <Tag text="Nft Collection" color="blue" />
+        </div>,
+        <div style={{ padding: '11px' }}>0x18...130e</div>,
+        <div style={{ padding: '11px' }}>
+            <PopoverDropdown
+                parent={<MoreVert fill="black" fontSize={32} />}
+                moveBody={-100}
+                children={[
+                    <PopoverElement
+                        key="0"
+                        height={50}
+                        width={260}
+                        text={'Testnet Server'}
+                        textSize={20}
+                        icon={<Testnet fill={color.white} fontSize={20} />}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Testnet Server')}
+                    />,
+                    <PopoverElement
+                        key="1"
+                        height={50}
+                        width={260}
+                        text={'Mainnet Server'}
+                        textSize={20}
+                        iconSize={30}
+                        icon={<Network fill={color.white} fontSize={20} />}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Mainnet Server')}
+                    />,
+                    <PopoverElement
+                        key="2"
+                        height={50}
+                        width={260}
+                        text={'Local Devchain Server'}
+                        textSize={20}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        icon={<Server fill={color.white} fontSize={20} />}
+                        textColor={color.white}
+                        onClick={() => alert('Local Devchain Server')}
+                    />,
+                ]}
+            />
+        </div>,
+    ],
+    [
+        <div style={{ padding: '11px' }}>
+            <Avatar theme="image" isRounded={true} size={36} />
+        </div>,
+        <div style={{ padding: '11px' }}>Moralis Magi</div>,
+        <div>
+            <Tag text="Nft Collection" color="blue" />
+        </div>,
+        <div style={{ padding: '11px' }}>0x18...130e</div>,
+        <div style={{ padding: '11px' }}>
+            <PopoverDropdown
+                parent={<MoreVert fill="black" fontSize={32} />}
+                moveBody={-100}
+                children={[
+                    <PopoverElement
+                        key="0"
+                        height={50}
+                        width={260}
+                        text={'Testnet Server'}
+                        textSize={20}
+                        icon={<Testnet fill={color.white} fontSize={20} />}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Testnet Server')}
+                    />,
+                    <PopoverElement
+                        key="1"
+                        height={50}
+                        width={260}
+                        text={'Mainnet Server'}
+                        textSize={20}
+                        iconSize={30}
+                        icon={<Network fill={color.white} fontSize={20} />}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Mainnet Server')}
+                    />,
+                    <PopoverElement
+                        key="2"
+                        height={50}
+                        width={260}
+                        text={'Local Devchain Server'}
+                        textSize={20}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        icon={<Server fill={color.white} fontSize={20} />}
+                        textColor={color.white}
+                        onClick={() => alert('Local Devchain Server')}
+                    />,
+                ]}
+            />
+        </div>,
+    ],
+    [
+        <div style={{ padding: '11px' }}>
+            <Avatar theme="image" isRounded={true} size={36} />
+        </div>,
+        <div style={{ padding: '11px' }}>Moralis Magi</div>,
+        <div>
+            <Tag text="Nft Collection" color="blue" />
+        </div>,
+        <div style={{ padding: '11px' }}>0x18...130e</div>,
+        <div style={{ padding: '11px' }}>
+            <PopoverDropdown
+                parent={<MoreVert fill="black" fontSize={32} />}
+                moveBody={-100}
+                children={[
+                    <PopoverElement
+                        key="0"
+                        height={50}
+                        width={260}
+                        text={'Testnet Server'}
+                        textSize={20}
+                        icon={<Testnet fill={color.white} fontSize={20} />}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Testnet Server')}
+                    />,
+                    <PopoverElement
+                        key="1"
+                        height={50}
+                        width={260}
+                        text={'Mainnet Server'}
+                        textSize={20}
+                        iconSize={30}
+                        icon={<Network fill={color.white} fontSize={20} />}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Mainnet Server')}
+                    />,
+                    <PopoverElement
+                        key="2"
+                        height={50}
+                        width={260}
+                        text={'Local Devchain Server'}
+                        textSize={20}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        icon={<Server fill={color.white} fontSize={20} />}
+                        textColor={color.white}
+                        onClick={() => alert('Local Devchain Server')}
+                    />,
+                ]}
+            />
+        </div>,
+    ],
+    [
+        <div style={{ padding: '11px' }}>
+            <Avatar theme="image" isRounded={true} size={36} />
+        </div>,
+        <div style={{ padding: '11px' }}>Moralis Magi</div>,
+        <div>
+            <Tag text="Nft Collection" color="blue" />
+        </div>,
+        <div style={{ padding: '11px' }}>0x18...130e</div>,
+        <div style={{ padding: '11px' }}>
+            <PopoverDropdown
+                parent={<MoreVert fill="black" fontSize={32} />}
+                moveBody={-100}
+                children={[
+                    <PopoverElement
+                        key="0"
+                        height={50}
+                        width={260}
+                        text={'Testnet Server'}
+                        textSize={20}
+                        icon={<Testnet fill={color.white} fontSize={20} />}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Testnet Server')}
+                    />,
+                    <PopoverElement
+                        key="1"
+                        height={50}
+                        width={260}
+                        text={'Mainnet Server'}
+                        textSize={20}
+                        iconSize={30}
+                        icon={<Network fill={color.white} fontSize={20} />}
+                        backgroundColor={'transparent'}
+                        textColor={color.white}
+                        onClick={() => alert('Mainnet Server')}
+                    />,
+                    <PopoverElement
+                        key="2"
+                        height={50}
+                        width={260}
+                        text={'Local Devchain Server'}
+                        textSize={20}
+                        iconSize={30}
+                        backgroundColor={'transparent'}
+                        icon={<Server fill={color.white} fontSize={20} />}
+                        textColor={color.white}
+                        onClick={() => alert('Local Devchain Server')}
+                    />,
+                ]}
+            />
+        </div>,
     ],
 ];